requirements.txt
appdirs==1.4.4
asgiref==3.3.1
bcrypt==3.1.7
blinker==1.4
certifi==2019.11.28
cffi==1.14.5
chardet==3.0.4
Click==7.0
colorama==0.4.3
cryptography==2.8
distlib==0.3.1
distro==1.4.0
Django==3.1.6
django-crontab==0.7.1
django-environ==0.4.5
djangorestframework==3.12.2
entrypoints==0.3
environ==1.0
fasteners==0.14.1
filelock==3.0.12
future==0.18.2
httplib2==0.14.0
idna==2.8
jeepney==0.6.0
keyring==18.0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==1.1.0
monotonic==1.5
oauthlib==3.1.0
olefile==0.46
paramiko==2.6.0
pbr==5.5.1
pexpect==4.6.0
Pillow==7.0.0
powerline-status==2.7
protobuf==3.6.1
psycopg2-binary==2.8.6
ptyprocess==0.7.0
pycparser==2.20
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyRFC3339==1.1
python-dateutil==2.7.3
pytz==2019.3
pyxdg==0.26
PyYAML==5.3.1
reportlab==3.5.34
requests==2.22.0
requests-unixsocket==0.2.0
SecretStorage==2.3.1
simplejson==3.16.0
six==1.14.0
sqlparse==0.4.1
stevedore==3.3.0
testresources==2.0.1
urllib3==1.25.8
virtualenv==20.4.2
virtualenv-clone==0.5.4
virtualenvwrapper==4.8.4
wadllib==1.3.3
